+++ extension/gloss-config.js
@@ -1,0 +1,1 @@
+globalThis.GLOSS_PAIRING_TOKEN = "";